Members delete and middleware validations created
