+++ routes/members.js
@@ -3,7 +3,7 @@ const express = require('express');
 const router = express.Router();
 
 const memberController = require('../controllers/memberController');
-const memberMiddleware = require('../validations/members');
+const memberMiddleware = require('../middlewares/membersMiddleware');
 
 router.get('/', memberController.readAll);
 
@@ -11,6 +11,6 @@ router.post('/', memberMiddleware.create, memberController.create);
 
 router.put('/:id', memberMiddleware.update, memberController.Update);
 
-router.delete('/softDeleteExample', memberController.softDelete);
+router.delete('/:id', memberMiddleware.delete, memberController.softDelete);
 
 module.exports = router;
